+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 * Добавлена поддержка сервисов 
   [AdImages](https://tech.yandex.ru/direct/doc/ref-v5/adimages/adimages-docpage/) 
   и [KeywordsResearch](https://tech.yandex.ru/direct/doc/ref-v5/keywordsresearch/keywordsresearch-docpage/)
+  
+* Удалена поддержка устаревших методов `CreateOfflineReport`, `DeleteOfflineReport`
+  и `GetOfflineReportList`.
 
 ## 4.0.1 [commit logs](https://github.com/biplane/yandex-direct/compare/4.0...4.0.1)